+++ Js2IL.Tests/BinaryOperator/ExecutionTests.cs
@@ -17,6 +17,9 @@ namespace Js2IL.Tests.BinaryOperator
         public Task BinaryOperator_AddNumberNumber() { var testName = nameof(BinaryOperator_AddNumberNumber); return ExecutionTest(testName); }
 
         [Fact]
+        public Task BinaryOperator_AddDynamicThenToNumber_StringPreserved() { var testName = nameof(BinaryOperator_AddDynamicThenToNumber_StringPreserved); return ExecutionTest(testName); }
+
+        [Fact]
         public Task BinaryOperator_AddObjectObject() { var testName = nameof(BinaryOperator_AddObjectObject); return ExecutionTest(testName); }
 
         [Fact]